Move Extension functions into separate files by type
